+++ src/main/java/org/testes/Matematica.java
@@ -9,11 +9,11 @@ public class Matematica extends Utilitario {
         this.nomeFuncao = nomeFuncao;
     }
 
-    public static double soma(double a, double b){
+    public static int soma(int a, int b){
         return a + b;
     }
 
-    public static double divisaoInteira(int a, int b){
+    public static double divisaoInteira(double a, double b){
         return a/b;
     }
 
@@ -21,7 +21,7 @@ public class Matematica extends Utilitario {
         return a - b;
     }
 
-    public static double multiplicacao(double a, double b){
+    public static double multiplicacao(int a, int b){
         return a * b;
     }
